fix form prepopulation semantics
